# docs/swagger.yaml
basePath: /
definitions:
...
  handyman.AddPortfolioRequest:
    description: Add portfolio item
    properties:
      caption:
        example: Remont lazienki
        type: string
      image_url:
        example: https://example.com/photo.jpg
        type: string
      sort_order:
        example: 0
        type: integer
    type: object
  handyman.CreatePricingRequest:
    description: Create pricing item
    properties:
      price_from:
        example: 100
        type: integer
      price_to:
        example: 200
        type: integer
      service_name:
        example: Wymiana kranu
        type: string
      sort_order:
        example: 0
        type: integer
      unit:
        example: per service
        type: string
    type: object
  handyman.FullProfileResponse:
    description: Full handyman profile with pricing and portfolio
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      bio:
        example: Hydraulik z 10-letnim doswiadczeniem
        type: string
      categories:
        items:
          type: string
        type: array
      company_name:
        example: Jan Kowalski Hydraulika
        type: string
      completion_pct:
        example: 85
        type: integer
      created_at:
        type: string
      districts:
        items:
          type: string
        type: array
      email:
        example: jan@example.com
        type: string
      emergency_available:
        example: false
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_available:
        example: true
        type: boolean
      is_verified:
        example: true
        type: boolean
      nip:
        example: "1234567890"
        type: string
      phone:
        example: "+48123456789"
        type: string
      portfolio:
        items:
          $ref: '#/definitions/handyman.PortfolioItemResponse'
        type: array
      pricing:
        items:
          $ref: '#/definitions/handyman.PricingItemResponse'
        type: array
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  handyman.PortfolioItemResponse:
    description: Portfolio item
    properties:
      caption:
        example: Remont lazienki
        type: string
      created_at:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image_url:
        example: https://example.com/photo.jpg
        type: string
      sort_order:
        example: 0
        type: integer
    type: object
  handyman.PricingItemResponse:
    description: Pricing list item
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      price_from:
        example: 100
        type: integer
      price_to:
        example: 200
        type: integer
      service_name:
        example: Wymiana kranu
        type: string
      sort_order:
        example: 0
        type: integer
      unit:
        example: per service
        type: string
    type: object
  handyman.ProfileListResponse:
    description: Paginated handyman profile list
    properties:
      has_more:
        example: true
        type: boolean
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      profiles:
        items:
          $ref: '#/definitions/handyman.ProfileResponse'
        type: array
      total:
        example: 50
        type: integer
    type: object
  handyman.ProfileResponse:
    description: Handyman profile information
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      bio:
        example: Hydraulik z 10-letnim doswiadczeniem
        type: string
      categories:
        items:
          type: string
        type: array
      company_name:
        example: Jan Kowalski Hydraulika
        type: string
      completion_pct:
        example: 85
        type: integer
      created_at:
        type: string
      districts:
        items:
          type: string
        type: array
      email:
        example: jan@example.com
        type: string
      emergency_available:
        example: false
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_available:
        example: true
        type: boolean
      is_verified:
        example: true
        type: boolean
      nip:
        example: "1234567890"
        type: string
      phone:
        example: "+48123456789"
        type: string
      updated_at:
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  handyman.UpdateProfileRequest:
    description: Update handyman profile
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      bio:
        example: Hydraulik z 10-letnim doswiadczeniem
        type: string
      categories:
        items:
          type: string
        type: array
      company_name:
        example: Jan Kowalski Hydraulika
        type: string
      districts:
        items:
          type: string
        type: array
      email:
        example: jan@example.com
        type: string
      emergency_available:
        example: false
        type: boolean
      is_available:
        example: true
        type: boolean
      nip:
        example: "1234567890"
        type: string
      phone:
        example: "+48123456789"
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      error:
        example: Invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  response.ValidationError:
    properties:
      error:
        example: Validation failed
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@fixapp.com
    name: API Support
  description: FixApp backend API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FixApp API
  version: "1.0"
paths:
...
  /handyman/portfolio:
    post:
      consumes:
      - application/json
      description: Adds a new photo to the handyman's portfolio
      parameters:
      - description: Portfolio item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handyman.AddPortfolioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handyman.PortfolioItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationError'
      security:
      - BearerAuth: []
      summary: Add portfolio photo
      tags:
      - handyman-profile
  /handyman/portfolio/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a photo from the handyman's portfolio
      parameters:
      - description: Portfolio item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Delete portfolio photo
      tags:
      - handyman-profile
  /handyman/pricing:
    post:
      consumes:
      - application/json
      description: Adds a new service pricing item to the handyman's profile
      parameters:
      - description: Pricing item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handyman.CreatePricingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handyman.PricingItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationError'
      security:
      - BearerAuth: []
      summary: Add pricing item
      tags:
      - handyman-profile
  /handyman/pricing/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a pricing item from the handyman's profile
      parameters:
      - description: Pricing item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Delete pricing item
      tags:
      - handyman-profile
  /handyman/profile:
    get:
      consumes:
      - application/json
      description: Returns the authenticated handyman's full profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handyman.FullProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get my handyman profile
      tags:
      - handyman-profile
    patch:
      consumes:
      - application/json
      description: Updates the authenticated handyman's profile
      parameters:
      - description: Profile update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handyman.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handyman.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationError'
      security:
      - BearerAuth: []
      summary: Update my handyman profile
      tags:
      - handyman-profile
  /handymen:
    get:
      consumes:
      - application/json
      description: Search for available handyman profiles with filters
      parameters:
      - description: Filter by category ID
        in: query
        name: category_id
        type: string
      - description: Filter by district ID
        in: query
        name: district_id
        type: string
      - description: Filter by verification status
        in: query
        name: verified
        type: boolean
      - description: Search by name or bio
        in: query
        name: search
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handyman.ProfileListResponse'
      summary: Search handymen
      tags:
      - handymen
  /handymen/{id}:
    get:
      consumes:
      - application/json
      description: Returns a handyman's full public profile with pricing and portfolio
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handyman.FullProfileResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get handyman public profile
      tags:
      - handymen
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"